requirements.txt
Pillow==10.3.0
google-generativeai==0.8.3
gradio==5.1.0
python-dotenv==1.0.1